Add Heikin-Ashi signal with logging, plotting, and MQL5 export

+++ src/trade_lab/optimization/param_space.py
@@ -83,8 +83,8 @@ class FloatParam:
     """
 
     name: str
-    low: Decimal
-    high: Decimal
+    low: float
+    high: float
     log: bool = False
